Worked on project structure
