+++ app/config.py
@@ -1,0 +1,8 @@
+class Config:
+    vanity_urls = {
+        "resin": "https://explorer.resincoin.ml",
+        "ssniper1": "https://ssniper1.ml",
+        "murka": "https://murkauserbot.ml",
+        "gay": "https://t.me/gdlbo",
+    }
+    expression = "^(https?:\/\/)?[a-zA-Z0-9_\-]+(\.[a-zAZ0-9_\-]+)+(\/.*)?$"
