+++ app/modules/git_tools.py
@@ -1,0 +1,5 @@
+import subprocess
+
+
+def get_git_revision_short_hash() -> str:
+    return subprocess.check_output(['git', 'rev-parse', '--short', 'HEAD']).decode('ascii').strip()
