+++ app/routes.py
@@ -1,41 +1,8 @@
-import sqlite3
-from flask import Flask, request, render_template, send_file, redirect
-import hashlib
 import re
-import subprocess
+import hashlib
 
-
-def get_git_revision_short_hash() -> str:
-    return subprocess.check_output(['git', 'rev-parse', '--short', 'HEAD']).decode('ascii').strip()
-
-
-expression = "^(https?:\/\/)?[a-zA-Z0-9_\-]+(\.[a-zAZ0-9_\-]+)+(\/.*)?$"
-
-vanity_urls = {
-    "resin": "https://explorer.resincoin.ml",
-    "ssniper1": "https://ssniper1.ml",
-    "murka": "https://murkauserbot.ml",
-    "gay": "https://t.me/gdlbo",
-}
-
-app = Flask(__name__)
-db = sqlite3.connect('URLs.db', check_same_thread=False)
-cur = db.cursor()
-for vanityurl in vanity_urls:
-        print(f"{vanityurl}:{vanity_urls[vanityurl]}")
-try:
-    cur.execute('CREATE TABLE URLs (url TEXT unique, hash TEXT PRIMARY KEY)')
-except:
-    pass
-
-for vanityurl in vanity_urls:
-    key = vanityurl
-    val = vanity_urls[vanityurl]
-    try:
-        cur.execute("INSERT INTO URLs (url,hash) VALUES (?,?)", (val, key,))
-        db.commit()
-    except:
-        pass
+from app import get_git_revision_short_hash, cur, db, config
+from app import app, render_template, send_file, request, redirect
 
 
 @app.route('/')
@@ -63,19 +30,20 @@ def shorten():
     if request.method == "GET":
         return redirect("/")
     url = request.form['url']
-    if not re.match(expression, url):
+    if not re.match(config.expression, url):
         return "Invalid URL"
-    hash = hashlib.sha512(url.encode('utf-8')).hexdigest()[:6]
+    url_hash = hashlib.sha512(url.encode('utf-8')).hexdigest()[:6]
     cur.execute("SELECT hash FROM URLs WHERE url=?", (url,))
     result = cur.fetchone()
     if result:
         return render_template("shorted.html", short=result[0])
     else:
-        cur.execute("INSERT INTO URLs (url,hash) VALUES (?,?)", (url, hash,))
+        cur.execute("INSERT INTO URLs (url,hash) VALUES (?,?)", (url, url_hash,))
         db.commit()
         cur.execute("SELECT * FROM URLs WHERE url=?", (url,))
         result = cur.fetchone()
         return render_template("shorted.html", short=result[1])
+
 
 @app.route("/<short>")
 def redirect_to_short(short):
@@ -85,6 +53,3 @@ def redirect_to_short(short):
         return redirect(url)
     else:
         return redirect(f"http://{url}")
-
-if __name__ == "__main__":
-    app.run(debug=True)
