+++ shorter.py
@@ -1,0 +1,4 @@
+from app import *
+
+if __name__ == '__main__':
+    app.run(debug=True)
